check existence of the dir before deletion

+++ scripts/stage-runner/stop.sh
@@ -2,12 +2,13 @@
 
 set -eu
 
-pushd .
-cd /home/ec2-user/RADAR-Docker/dcompose-stack/radar-cp-hadoop-stack
-./bin/radar-docker down
-popd
-
-rm -rf /home/ec2-user/RADAR-Docker
+if [ -d "/home/ec2-user/RADAR-Docker" ]; then
+    pushd .
+    cd /home/ec2-user/RADAR-Docker/dcompose-stack/radar-cp-hadoop-stack
+    ./bin/radar-docker down
+    popd
+    rm -rf /home/ec2-user/RADAR-Docker
+fi
 
 # Remove nginx if installed to free port 80
 systemctl stop nginx
